chg: [core] AlgoConf splitting in its own file
